refactor: moved libraries to allow SSG
