+++ apps/web/app/[lng]/layout.next.tsx
@@ -1,16 +1,12 @@
-'use client'
-
 import { dir } from 'i18next'
-import { usePathname, useRouter } from 'next/navigation'
 import Script from 'next/script'
-import { useEffect } from 'react'
 
-import { LanguageContextProvider, useDefaultLanguage } from '@sdlgr/i18n-client'
+import { LanguageContextProvider } from '@sdlgr/i18n-tools'
 
 import { MainLayout } from '@web/components/MainLayout/MainLayout'
 import StyledComponentsRegistry from '@web/components/StyledComponentsRegistry/StyledComponentsRegistry'
 import { getServerTranslation } from '@web/i18n/server'
-import { fallbackLng, languages } from '@web/i18n/settings'
+import { languages } from '@web/i18n/settings'
 
 interface RouteProps {
   params: {
@@ -32,11 +28,11 @@ export async function generateMetadata({ params }: GenerateMetadataProps) {
     colorScheme: 'dark',
     metadataBase: new URL('https://www.sawl.dev'),
     alternates: {
-      canonical: '/es',
+      canonical: '/en',
       languages: {
-        'es-ES': '/es',
         'en-US': '/en',
         'en-UK': '/en',
+        'es-ES': '/es',
       },
     },
   }
@@ -50,22 +46,6 @@ export default function RootLayout({
   children,
   params: { lng },
 }: RootLayoutProps) {
-  const pathname = usePathname()
-  const router = useRouter()
-  const defaultLanguage = useDefaultLanguage({ languages, fallbackLng })
-
-  // if (!languages.some((loc) => lng === loc)) {
-  //   redirect(`/${defaultLanguage}/${differencePath}`)
-  // }
-
-  useEffect(() => {
-    if (!languages.some((loc) => lng === loc)) {
-      const urlParts = pathname.split('/').filter(Boolean)
-      const differencePath = urlParts.slice(1, urlParts.length).join('/')
-      router.push(`/${defaultLanguage}/${differencePath}`)
-    }
-  }, [defaultLanguage, lng, pathname, router])
-
   const googleAnalyticsId = process.env.GOOGLE_ANALYTICS_ID
   return (
     <html lang={lng} dir={dir(lng)} data-testid="root-html">
