+++ apps/web/components/MainLayout/MainLayout.tsx
@@ -1,5 +1,12 @@
 'use client'
 
+import { usePathname, useRouter } from 'next/navigation'
+import { useContext, useEffect } from 'react'
+
+import { LanguageContext, useDefaultLanguage } from '@sdlgr/i18n-tools'
+
+import { fallbackLng, languages } from '@web/i18n/settings'
+
 import { StyledContent, StyledPage } from './MainLayout.styles'
 import { CookieBanner } from './components/CookieBanner/CookieBanner'
 import { Footer } from './components/Footer/Footer'
@@ -11,6 +18,19 @@ interface LayoutProps {
 }
 
 export function MainLayout({ children }: LayoutProps) {
+  const { language } = useContext(LanguageContext)
+  const pathname = usePathname()
+  const router = useRouter()
+  const defaultLanguage = useDefaultLanguage({ languages, fallbackLng })
+
+  useEffect(() => {
+    if (!languages.some((loc) => language === loc)) {
+      const urlParts = pathname.split('/').filter(Boolean)
+      const differencePath = urlParts.slice(1, urlParts.length).join('/')
+      router.push(`/${defaultLanguage}/${differencePath}`)
+    }
+  }, [defaultLanguage, language, pathname, router])
+
   return (
     <StyledPage>
       <ProgressBar />
