+++ libs/i18n-client/src/index.ts
@@ -1,3 +1,1 @@
 export * from './lib/i18n-client'
-export * from './lib/LanguageProvider/LanguageProvider'
-export * from './lib/useDefaultLanguage/useDefaultLanguage'
