+++ libs/i18n-tools/README.md
@@ -1,0 +1,7 @@
+# i18n-tools
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test i18n-tools` to execute the unit tests via [Jest](https://jestjs.io).
