+++ tsconfig.base.json
@@ -24,6 +24,7 @@
       "@sdlgr/i18n-client": ["libs/i18n-client/src/index.ts"],
       "@sdlgr/i18n-config": ["libs/i18n-config/src/index.ts"],
       "@sdlgr/i18n-server": ["libs/i18n-server/src/index.ts"],
+      "@sdlgr/i18n-tools": ["libs/i18n-tools/src/index.ts"],
       "@sdlgr/link": ["libs/link/src/index.ts"],
       "@sdlgr/main-theme": ["libs/main-theme/src/index.ts"],
       "@sdlgr/storage": ["libs/storage/src/index.ts"],
